table不涉及函数调用的部分指令

+++ LuaVM.Net/Core/VM/Api/Internals/String.cs
@@ -4,7 +4,7 @@ namespace LuaVM.Net.Core
 {
     internal static class String
     {
-        internal static int Len(LuaValue value)
+        internal static long Len(LuaValue value)
         {
             var s = value.GetString();
             return s.Length;
